refactor: split burrow.go into themed root files + add smoke tests
Mirrors den 0.16's pattern: aliases.go holds type aliases, themed
wrapper files (server, config, web, context, tasks, pagination) replace
the 510-line single facade. New burrow_test.go gives the root package
smoke coverage instead of [no test files].

No API change.
