+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to Burrow are documented here. The format is based on [Keep 
 - **`contrib/jobs` docs corrected to "Den-backed (SQLite + Postgres)".** Doc comments and READMEs called it "SQLite-backed" — that was true in v0.1 but the repository has been Postgres-aware since Den arrived. Code behaviour unchanged.
 - **`burrow` package split into themed public sub-packages.** Implementation moves into `burrow/app` (App interface, capability interfaces, AppConfig, NavItem, Config and sub-configs, context helpers), `burrow/server` (Server, boot sequence, templates, listener, TLS, OpenDB), `burrow/web` (HandlerFunc, HTTPError, Handle, JSON, Render, Bind, Validate), `burrow/tasks` (Queue, Task, Result, JobConfig), `burrow/pagination` (PageRequest, PageURL, OffsetResult). The burrow root retains type aliases and wrapper functions for the everyday API — `burrow.App`, `burrow.AppConfig`, `burrow.HasRoutes`, `burrow.NewServer`, `burrow.Handle`, `burrow.JSON`, `burrow.Render`, `burrow.Bind`, `burrow.Validate`, `burrow.DefineTask`, `burrow.ParsePageRequest` and so on continue to work via aliases. Less-frequent operations are also reachable via direct sub-package import (`tasks.DefineTask[P]`, `pagination.ParsePageRequest`, `app.Config` sub-configs, `web.HandlerFunc`). `Config.resolvedTLSMode` is now exported as `ResolvedTLSMode` because `tls.go` lives in `burrow/server` and calls it across the package boundary. No behaviour changes.
 - **Den bumped to v0.16.0.** Internal refactor in Den (its `internal/core` split into public sub-packages); the `den.X` alias surface burrow uses is API-compatible — no changes required.
+- **`burrow.go` split into themed root files for readability**, with smoke tests added to the root package. No API change.
 
 ## 0.23.0 — 2026-05-21
